assignment-10/backend/requirements-test.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic==2.11.7
pydantic-settings==2.10.1
pytest>=7.0.0
pytest-cov>=4.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
httpx>=0.24.0
coverage>=7.0.0 